fix(4.5a): fast-gate 27396114894 双红修 — pcm_hook 单测 merge-window stub + pageno 无符号移位
run 27396114894 两 job 红,均为 9-commit 首推才暴露(本机旧二进制掩盖):

1. macOS link-smoke: test_cluster_bufmgr_pcm_hook 引 bufpage.h
   PageSetLSN inline(§3.3b merge-window 钩),但独立二进制不链
   cluster_recovery_merge.o → cluster_recmerge_* 4 全局未定义。
   补 inactive stub(窗门恒 false)。本机此前 sweep 一直吃旧二进制,
   清光重编 91 个全过后才复现 — 单测 Makefile 依赖未列 bufpage.h

2. Security/cppcheck: cluster_remote_xact_pageno 有符号左移
   (shiftNegativeLHS portability,baseline-diff strict 拦截)。
   origin_node 调用方已验非负,改 uint32 移位消 UB

本地: cppcheck 0 findings(匹配 baseline)/ check-format 0 /
cluster_unit 91 binaries 清重编全过 / t/248 54/54

Spec: spec-4.5a-shared-storage-data-backend.md

+++ src/include/cluster/cluster_remote_xact.h
@@ -73,8 +73,10 @@ struct XLogReaderState; /* forward; avoid xlogreader.h in this header */
 static inline int
 cluster_remote_xact_pageno(int origin_node, TransactionId xid)
 {
-	return (origin_node << CLUSTER_REMOTE_XACT_ORIGIN_PAGE_SHIFT)
-		   | (int)(xid / CLUSTER_REMOTE_XACT_ENTRIES_PER_PAGE);
+	/* Unsigned shift: origin_node is validated non-negative by every
+	 * caller, but a signed LHS is technically UB (cppcheck portability). */
+	return (int)(((uint32)origin_node << CLUSTER_REMOTE_XACT_ORIGIN_PAGE_SHIFT)
+				 | (uint32)(xid / CLUSTER_REMOTE_XACT_ENTRIES_PER_PAGE));
 }
 
 static inline int
